+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed glitchy transitions.
 - Added support for capitalization-based address checksums.
 - Send value is no longer limited by javascript number precision, and is always in ETH.
+- Added ability to generate new accounts.
 
 ## 1.8.4 2016-05-13